% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_corpus.R
\name{clean_corpus}
\alias{clean_corpus}
\title{Clean Corpus}
\usage{
clean_corpus(ipath, odir, ncores, clean_commands_str)
}
\arguments{
\item{ipath}{A string specifying the path to all the text files to handle.}

\item{odir}{A string specifying the path to an output directory.}

\item{ncores}{A number specifying the number of cores to use.}

\item{clean_commands_str}{A string containing the combined commands for the cleaning script.}
}
\value{
A character vector of all the files taht were cleaned.
}
\description{
A function that cleans a corpus based on user specification. 
Handles each file in the ipath in parallel and runs clean_file on each file.
Outputs the cleaned version of the file into the output directory specified. 
Make sure output directory either doesn't exist (yet) or has nothing important in it,
As this function will delete whatever is already in there.
Look at the documentation for clean_file to see the commands to pass to the cleaning 
script.
}
\examples{
\dontrun{
clean_corpus("/path/to/corpus/", "./cleaned/", 20, "-lnprsd --maintain-newlines --min-size 2")
}
}
