% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_corpus.R
\name{clean_file}
\alias{clean_file}
\title{Clean File}
\usage{
clean_file(ifile, odir, clean_commands_str)
}
\arguments{
\item{ifile}{A string containing the path to the input file.}

\item{odir}{A string containing the path to the output directory.}

\item{clean_commands_str}{A string containing the combined commands for the cleaning script.}
}
\value{
A string containing the name of the file that was cleaned.
}
\description{
A wrapper around the clean.py script in /inst/python/.
Specify the path to file you want to clean, an output directory and a string 
containing the commands you want to send to the cleaning script.
Be sure to pass the maintain-newlines parameter if your files are in a format 
where many documents are in one text file delimited by newlines.
Also, if you plan on using the lemmatizing functions be sure to add the --lemma flag
here as well, as it prepares the corpus for tree tagger lemmatizing.
For conveniance I am putting all the possible commands that can be passed to that python
script here. \cr 
     -l                     : if words should be lowercased \cr
     -n                     : if digits should be stripped \cr
     -p                     : if punctuation should be stripped \cr
     -r                     : if roman numerals should be stripped \cr
     -s                     : if stop words should be stripped \cr
     -d                     : if non dictionary words should be stripped \cr
     -t                     : if tweet specific cleaning options should be used \cr
     --additional           : if triggered then adds all stopwords and dictionary files \cr
     --lemma                : if triggered convert common regexes to common form \cr
     --no-usernames         : if usernames (ampersand<name>) should be removed \cr
     --maintain-newlines    : use space for deilm instead of default(newline) \cr
     --min-size [N]         : specify the minimum size for a token (default=2)  \cr
}
\examples{
\dontrun{
clean_file("myfile.txt", "./cleaned/", "-d")
clean_file("myfile.txt", "./cleaned/", "-lnp")
clean_file("myfile.txt", "./cleaned/", "-lnprsdt")
clean_file("myfile.txt", "./cleaned/", "-lnprsdt --additional")
clean_file("myfile.txt", "./cleaned/", "-lnprsdt --lemma --maintain-newlines --min-size 3")
}
}
