% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dos2unix}
\alias{dos2unix}
\title{Dos to Unix line endings}
\usage{
dos2unix(ifilename, ofilename)
}
\arguments{
\item{ifilename}{A string specifying path to input file.}

\item{ofilename}{A string specifying path to output file.}
}
\description{
Removes the windows dos \\r from input file.
}
