% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_corpus.R
\name{filter_corpus}
\alias{filter_corpus}
\title{Filter Corpus}
\usage{
filter_corpus(words, ipath, ncores)
}
\arguments{
\item{words}{A character vector of all the words to remove.}

\item{ipath}{A string specifying the path to all the text files to handle.}

\item{ncores}{A number specifying the number of cores to use.}
}
\description{
A function that removes all occurances from a given list of words from each (.txt) file in ipath.
Generally used after some analysis of the summary_corpus results. To remove the most and least frequent
terms. Handles each file in the ipath in parallel over the number of cores specified. Runs filter_file
on each file in the ipath.
}
\examples{
\dontrun{
filter_corpus(most_common_terms, "/path/to/corpus/", 10)
filter_corpus(stopwords, "/path/to/corpus/", 10)
}
}
