% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_corpus.R
\name{filter_file}
\alias{filter_file}
\title{Filter File}
\usage{
filter_file(words, ifilepath)
}
\arguments{
\item{words}{A character vector of all the words to remove.}

\item{ifilepath}{A string specifying the path to the input file.}
}
\value{
ifilepath
}
\description{
A wrapper around the rcpp_filter function.
Given a file and a list of words. Remove each occurance of those words from that file.
Will modify the input file, so no output file is specified.
}
\examples{
\dontrun{
filter_file(most_common_terms, "/path/to/file.txt")
filter_file(stopwords, "/path/to/file.txt")
}
}
