% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_abundant}
\alias{get_abundant}
\title{Get List of Abundant Terms}
\usage{
get_abundant(wf, ndocs, abundance)
}
\arguments{
\item{wf}{A data table containing the word and document frequencies accross the corpus.}

\item{ndocs}{A number specifying the total number of unique documents in the corpus.}

\item{abundance}{A number either decimal or whole; interpreted as percent, whole as count.}
}
\value{
words A character vector of all the abundant terms.
}
\description{
A function to analyze the output of the summary_corpus similar to get_spare.
Returns words that appeared in more than or equal to X percent of documents, 
if you pass X as a decimal. Otherwise, if X is a whole number
returns the words that appeared in X or more documents.
}
\examples{
\dontrun{
sparse = get_abundant(wf, 100, .95)
sparse = get_abundant(wf, 100, 95) 
}
}
