% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_sparse}
\alias{get_sparse}
\title{Get List of Sparse Terms}
\usage{
get_sparse(wf, ndocs, sparsity)
}
\arguments{
\item{wf}{A data table containing the word and document frequencies accross the corpus.}

\item{ndocs}{A number specifying the total number of unique documents in the corpus.}

\item{sparsity}{A number either decimal or whole; interpreted as percent, whole as count.}
}
\value{
words A character vector of all the sparse terms.
}
\description{
A function to analyze the output of the summary_corpus. Returns words that appeared in less
than or equal to X percent of documents, if you pass X as a decimal. Otherwise, if X is a whole number
returns the words that appeared in X or less documents.
}
\examples{
\dontrun{
sparse = get_sparse(wf, 100, .03)
sparse = get_sparse(wf, 100, 3)
}
}
