% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_top_terms}
\alias{get_top_terms}
\title{Get List of Most Frequent Terms}
\usage{
get_top_terms(wf, nterms, count)
}
\arguments{
\item{wf}{A data table containing the word and document frequencies accross the corpus.}

\item{nterms}{A number specifying the total number of unique words in the corpus.}

\item{count}{A number either decimal or whole; interpreted as percent, whole as count.}
}
\value{
words A character vector of the least frequent terms
}
\description{
Similar to get_sparse but looks at word frequency not doc count.
If X is whole number, returns the X most frequent terms. If X is decimal
returns the X% most frequent words.
}
\examples{
\dontrun{
infreq = get_top_terms(wf, 100000, 5000) #returns 5000 most common terms
infreq = get_top_terms(wf, 100000, .05) #returns the top 5\% of terms
}
}
