% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemma_corpus.R
\name{lemma_corpus}
\alias{lemma_corpus}
\title{Clean Corpus}
\usage{
lemma_corpus(ipath, odir, ncores,
  cmd = "/opt/tree-tagger/bin/tree-tagger",
  param = "/opt/tree-tagger/lib/english.par")
}
\arguments{
\item{ipath}{A string specifying the path to the input directory with all the text files to lemma.}

\item{odir}{A string specifying the output directory path.}

\item{ncores}{A number specifying the number of cores ot use.}

\item{cmd}{**optional** path to the tree taggery binary on your system.}

\item{param}{**optional** path to the param file to use.}
}
\value{
A character vector of all the files that were lemmatized.
}
\description{
A function that lemmatizes a corpus using tree-tagger.
Handles each file in the ipath in parallel and runs lemma_file on each file.
Outputs to the output directory specified. Make sure the output directory 
doesn't exist already or has nothing in there.
}
\examples{
\dontrun{
lemma_corpus("/path/to/corpus/", "./lemmad/", 20)
lemma_corpus("corpus/", "lemmad/", 20, cmd="~/tt/bin/tree-tagger", param="~/tt/lib/english.par")
}
}
