% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemma_corpus.R
\name{lemma_file}
\alias{lemma_file}
\title{Lemma File}
\usage{
lemma_file(ifile, odir, cmd = "/opt/tree-tagger/bin/tree-tagger",
  param = "/opt/tree-tagger/lib/english.par")
}
\arguments{
\item{ifile}{A string containing the path to the input file.}

\item{odir}{A string containign the path to the output file.}

\item{cmd}{**optional** path to the tree taggery binary on your system.}

\item{param}{**optional** path to the param file to use.}
}
\value{
A string containing the name of the file that was lemmatized.
}
\description{
System call to tree-tagger binary to lemmatize file based on our tokenization.
Give input filename and outputfilename, the function will handle creating
temporary files, and the pre and post processing of the tree-tagger lemmatization necessary.
}
\examples{
\dontrun{
lemma_file("ifile.txt", "ofile.txt")
lemma_file("ifile.txt", "ofile.txt", cmd="~/tt/bin/tree-tagger", param="~/tt/lib/english.par")
}
}
