% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_doccount}
\alias{rcpp_doccount}
\title{Rcpp Doccount}
\usage{
rcpp_doccount(ipath, fileflag)
}
\arguments{
\item{ipath}{A string specifying the path to the input file.}

\item{fileflag}{An int, set to 1 if only one document per file, 0 if each document is on a newline.
returns An int vector of the results.}
}
\description{
Given a file returns the number of words for each document in the file.
Assumes documents are delimited
by newlines. If the file contains only one document then set fileflag arg to 1.
}
