% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_filter}
\alias{rcpp_filter}
\title{Rcpp Filter}
\usage{
rcpp_filter(words, ifilename)
}
\arguments{
\item{words}{A string vector containing the list of words to be removed
from the file.}

\item{ifilename}{A string containing the path to the inputfile>
Returns the name of the file that was modified.}
}
\description{
Given a vector of words and an input file, search through the file
and remove all instances of that word. Overwrites the input file
with the modified content.
}
