% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_split}
\alias{rcpp_split}
\title{Rcpp Split}
\usage{
rcpp_split(fpath, odir, splitter, count)
}
\arguments{
\item{fpath}{A string specifying the path to the input file.}

\item{odir}{A string specifying the path to the output.}

\item{splitter}{Either l or c, 'l' for lines, 'c' for kilobytes.}

\item{count}{Number of lines or kilobytes to split the file on based on splitter.
Returns number of output files.}
}
\description{
This function reads in a file and splits it into smaller segments
either by byte size or by line count dependent on user input.
When finished, it deletes the original file.
}
