% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_summary}
\alias{rcpp_summary}
\title{Rcpp Summary}
\usage{
rcpp_summary(ipath, fileflag)
}
\arguments{
\item{ipath}{A string specifying the path to the input file.}

\item{fileflag}{An int, set to 1 if only one document per file, 0 if each document is on a newline.
returns A string vector of the results.}
}
\description{
Given a file returns each unique word, the number of times that word appeared,
and the number of documents that word appeared in. Assumes documents are delimited
by newlines. If the file contains only one document then set delim arg to 1.
}
