% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_corpus.R
\name{summary_corpus}
\alias{summary_corpus}
\title{Summary of Corpus}
\usage{
summary_corpus(ipath, ncores, flag = 0)
}
\arguments{
\item{ipath}{A string specifying the path to the input files.}

\item{ncores}{A number specifying the number of cores to use.}

\item{flag}{**optional** A number specifying if documents are delimited by newline (set to 0)
or each document is in a different text file.}
}
\value{
A dataframe object that has merged the dataframes for each file.
	   Has term,freq,doccount for each term.
}
\description{
A function that calculates word frequency and document frequency
for all the words in the corpus. The output can then be analyzed
to remove outlier words, or stop words. Handles each file in 
parallel over the number of cores specified using parlapply. 
Runs summary_file function on each of the files in the ipath.
}
\examples{
\dontrun{
summary_corpus("/path/to/corpus/", 0)
}
}
