% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_corpus.R
\name{clean_file}
\alias{clean_file}
\title{Clean File}
\usage{
clean_file(ifile, odir, clean_commands_str)
}
\arguments{
\item{ifile}{A string containing the path to the input file.}

\item{odir}{A string containing the path to the output directory.}

\item{clean_commands_str}{A string containing the combined commands for the cleaning script.}
}
\value{
A string containing the name of the file that was cleaned.
}
\description{
A wrapper around the clean.py script in /inst/python/.
Specify the path to file you want to clean, an output directory and a string 
containing the commands you want to send to the cleaning script.
Be sure to pass the maintain-newlines parameter if your files are in a format 
where many documents are in one text file delimited by newlines.
Also, if you plan on using the lemmatizing functions be sure to add the --lemma flag
here as well, as it prepares the corpus for tree tagger lemmatizing.
For conveniance I am putting all the possible commands that can be passed to that python
script here. \cr 
		 \tabular{rll}{
        -l                    \tab    :  if words should be lowercased \cr
        -n                    \tab    :  if digits should be stripped \cr
        -p                    \tab    :  if punctuation should be stripped \cr
        -r                    \tab    :  if roman numerals should be stripped \cr
        -s                    \tab    :  if stop words should be stripped \cr
        -d                    \tab    :  if non dictionary words should be stripped \cr
        -t                    \tab    :  if tweet specific cleaning options should be used \cr
        ----additional        \tab    :  adds all stopwords and dictionary files \cr
        ----lemma             \tab    :  convert common regexes to common form \cr
        ----no-usernames      \tab    :  remove twitter usternames ampersand<name> \cr
        ----maintain-newlines \tab    :  use space for delim instead of default (newline) \cr
        ----min-size [N]      \tab    :  specify the minimum size for a token (default=2)  \cr
	}
}
\examples{
\dontrun{
clean_file("myfile.txt", "./cleaned/", "-d")
clean_file("myfile.txt", "./cleaned/", "-lnp")
clean_file("myfile.txt", "./cleaned/", "-lnprsdt")
clean_file("myfile.txt", "./cleaned/", "-lnprsdt --additional")
clean_file("myfile.txt", "./cleaned/", "-lnprsdt --lemma --maintain-newlines --min-size 3")
}
}
